Move Pagination into full component
Moves pagination component out of ResultList
Refactor `usePagination` hook
