+++ src/components/Pagination/index.ts
@@ -1,0 +1,2 @@
+export * from './Pagination';
+export * from './usePagination';
