+++ src/components/index.ts
@@ -8,10 +8,11 @@ export * from './Footer';
 export * from './images';
 export * from './Layout';
 export * from './ListBox';
+export * from './Metatags';
 export * from './Metrics';
 export * from './NavBar';
 export * from './NumFound';
-export * from './Metatags';
+export * from './Pagination';
 export * from './ResultList';
 export * from './SearchBar';
 export * from './SearchExamples';
